:sparkles: Add flavor for android
