+++ android/app/.gitignore
@@ -1,0 +1,1 @@
+**/google-services.json
