+++ android/app/src/dev/AndroidManifest.xml
