+++ flutter_launcher_icons-dev.yaml
@@ -1,0 +1,4 @@
+flutter_icons:
+  image_path: "assets/icons/dev.png"
+  android: true
+  ios: true
