+++ flutter_launcher_icons-prod.yaml
@@ -1,0 +1,4 @@
+flutter_icons:
+  image_path: "assets/icons/release.png"
+  android: true
+  ios: true
